# internal/api/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flagcel API
  description: Feature flag service with CEL-based targeting rules.
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
...
paths:
...
  /environments/{environment_id}/flags/{key}/rules:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/FlagKey"
    get:
      tags: [Rules]
      summary: List rules for a flag in an environment
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleListEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
    post:
      tags: [Rules]
      summary: Create a rule in an environment
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "200":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
  /environments/{environment_id}/flags/{key}/rules/reorder:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/FlagKey"
    post:
      tags: [Rules]
      summary: Reorder rules in an environment
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRulesRequest"
      responses:
        "204":
          description: Rules reordered
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
  /environments/{environment_id}/flags/{key}/rules/{id}:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/FlagKey"
      - $ref: "#/components/parameters/RuleID"
    get:
      tags: [Rules]
      summary: Get a rule in an environment
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
    put:
      tags: [Rules]
      summary: Update a rule in an environment
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
    delete:
      tags: [Rules]
      summary: Delete a rule in an environment
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Rule deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
  /flags/{key}/rules:
    parameters:
      - $ref: "#/components/parameters/FlagKey"
    get:
      tags: [Rules]
      summary: List rules for a flag
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleListEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Rules]
      summary: Create a rule
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "200":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
  /flags/{key}/rules/reorder:
    parameters:
      - $ref: "#/components/parameters/FlagKey"
    post:
      tags: [Rules]
      summary: Reorder rules
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRulesRequest"
      responses:
        "204":
          description: Rules reordered
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
  /flags/{key}/rules/{id}:
    parameters:
      - $ref: "#/components/parameters/FlagKey"
      - $ref: "#/components/parameters/RuleID"
    get:
      tags: [Rules]
      summary: Get a rule
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [Rules]
      summary: Update a rule
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Rules]
      summary: Delete a rule
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Rule deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RuleNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: flagcel_session
    bearerAPIKey:
      type: http
      scheme: bearer
  parameters:
    EnvironmentID:
      name: environment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FlagKey:
      name: key
      in: path
      required: true
      schema:
        type: string
        example: new-checkout
    RuleID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [message, data]
      properties:
        message:
          type: string
          example: success
        data: {}
...
    Rollout:
      type: object
      required: [percentage]
      properties:
        percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 25
        bucket_by:
          type: string
          example: user.id
    Rule:
      type: object
      required: [id, expression, rollout, value, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        description: { type: string }
        expression:
          type: string
          example: "user.country == 'US'"
        rollout:
          $ref: "#/components/schemas/Rollout"
        value:
          $ref: "#/components/schemas/FlagValue"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, nullable: true }
        deleted_by: { type: string, format: uuid, nullable: true }
...
    CreateRuleRequest:
      type: object
      required: [expression, rollout]
      properties:
        description: { type: string }
        expression: { type: string }
        rollout:
          $ref: "#/components/schemas/Rollout"
        value:
          allOf:
            - $ref: "#/components/schemas/FlagValue"
          default: true
    UpdateRuleRequest:
      $ref: "#/components/schemas/CreateRuleRequest"
    ReorderRulesRequest:
      type: object
      required: [rule_ids]
      properties:
        rule_ids:
          type: array
          items:
            type: string
            format: uuid
...
    FlagValue:
      nullable: true
      oneOf:
        - type: boolean
        - type: string
        - type: number
        - type: object
          additionalProperties: true
        - type: array
          items: {}
...
    RuleEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Rule"
    RuleListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Rule"
...
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - FLAG_NOT_FOUND
            - RULE_NOT_FOUND
            - ENVIRONMENT_NOT_FOUND
            - ENVIRONMENT_KEY_TAKEN
            - ENVIRONMENT_IN_USE
            - DEFAULT_ENVIRONMENT
            - CONTEXT_NOT_FOUND
            - CONTEXT_NAME_TAKEN
            - API_KEY_NOT_FOUND
            - INVALID_REQUEST
            - BAD_REQUEST
            - AUTH_NOT_CONFIGURED
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - FORBIDDEN
            - INTERNAL_ERROR
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    FlagNotFound:
      description: Flag not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RuleNotFound:
      description: Rule not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidRequest:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"